fix(frontend): satisfy ESLint react-refresh and exhaustive-deps
Move ProjectDetails helpers to separate modules so index only exports the page. Wrap company registration setBasicInfo/setContactInfo in useCallback for stable useEffect deps.
